Fixes weird extra lib test directories that were accidentally created on ship_feature branch.
