+++ lib/ship.rb
